requirements.txt
streamlit
fastapi
uvicorn[standard]
pyyaml
pydantic